proj3: race conditions:
I fixed the first nasty race condition:
The problem was that my code needed the controller to be
the first process to connect to each branch. This was not
always happening. To solve this, I make the controller send
each branch a dumby marker message before the branches are
allowed to connect to each other