proj4: fixed connection recovery
So I did something sort of dumb, but it works, so I'm keeping it.
The connections are not persistent, meaning that every time that we do
an RPC, we open the connection, do the RPC, and then close the
connection. This is not efficient, but it makes connecting to a down
node easier.